+++ samples/sampler/stories/qa/VirtualGridList.js
@@ -36,6 +36,7 @@ const
 				{...rest}
 				label={subText}
 				src={source}
+				style={{margin: 0}}
 			>
 				{text}
 			</ImageItem>
@@ -172,8 +173,8 @@ storiesOf('VirtualGridList', module)
 				horizontalScrollbar={select('horizontalScrollbar', prop.scrollbarOption, Config)}
 				itemRenderer={renderItem}
 				itemSize={{
-					minWidth: ri.scale(number('minWidth', Config, 768)),
-					minHeight: ri.scale(number('minHeight', Config, 588))
+					minWidth: ri.scale(number('minWidth', Config, 688)),
+					minHeight: ri.scale(number('minHeight', Config, 570))
 				}}
 				key={select('scrollMode', prop.scrollModeOption, Config)}
 				noScrollByWheel={boolean('noScrollByWheel', Config)}
@@ -181,7 +182,7 @@ storiesOf('VirtualGridList', module)
 				onScrollStart={action('onScrollStart')}
 				onScrollStop={action('onScrollStop')}
 				scrollMode={select('scrollMode', prop.scrollModeOption, Config)}
-				spacing={ri.scale(number('spacing', Config, -48))}
+				spacing={ri.scale(number('spacing', Config, 24))}
 				spotlightDisabled={boolean('spotlightDisabled', Config, false)}
 				verticalScrollbar={select('verticalScrollbar', prop.scrollbarOption, Config)}
 				wrap={wrapOption[select('wrap', ['false', 'true', '"noAnimation"'], Config)]}