+++ CHANGELOG.md
@@ -10,10 +10,21 @@ The following is a curated list of changes in the Enact sandstone module, newest
 
 ### Added
 
+- `sandstone/Scroller` prop `fadeOut` to show fade-out effect
+- `sandstone/VirtualList.VirtualGridList` prop `noAffordance` to remove affordance when scrolling by keys
+- `sandstone/Item` prop `centered`
 - `sandstone/Panels` and `sandstone/Panels.WizardPanel` props `onTransition` and `onWillTransition`
 - `sandstone/Panels.WizardPanel` prop `noAnimation` to suppress view transition animation
 - `sandstone/Slider` and `sandstone/ProgressBar` prop `showAnchor` to display anchor based on `progressAnchor` value
 - `sandstone/Scroller` and `sandstone/VirtualList` prop `initialHiddenHeight` to reduce a vertical scroll height being enlarged when collapsing a header
+
+### Changed
+
+- `sandstone/ImageItem` focus effect of vertical image item with labels
+
+### Fixed
+
+- `sandstone/Picker` horizontal joined height in large text mode
 
 ## [1.0.0-alpha.9] - 2020-04-20